Fix sudden accelration error in Auto mode Fix getVf error in SRampGenerator.cpp Add setting initial velocity to 0 in Car.cp

+++ lib/SRampGenerator.cpp
@@ -120,7 +120,7 @@ float SRampGenerator::getVf(void)
 {
     if (index <= totalTimeFrames)
     {
-        lastV = (int16_t)(commandV[index++]);
+        lastV = commandV[index++];
         return lastV;
     }
     return commandV[totalTimeFrames];
